backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.31,<3.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
google-auth>=2.29.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
